+++ apps/controllerx/cx_core/integration/state.py
@@ -18,6 +18,11 @@ class StateIntegration(Integration):
         )
 
     async def state_callback(
-        self, entity: Optional[str], attribute: Optional[str], old, new, kwargs
+        self,
+        entity: Optional[str],
+        attribute: Optional[str],
+        old: Optional[str],
+        new: str,
+        kwargs,
     ) -> None:
-        await self.controller.handle_action(new)
+        await self.controller.handle_action(new, previous_state=old)